backend/requirements.txt
fastapi==0.115.8
langchain==0.3.17
langchain_core==0.3.33
langchain_groq==0.2.4
langchain_huggingface==0.1.2
langgraph==0.2.69
motor==3.7.0
pydantic==2.10.6
pydantic_settings==2.7.1
qdrant_client==1.13.2
tavily_python==0.5.0
typing_extensions==4.12.2
uvicorn==0.34.0
